+++ src/chathpc/app/__main__.py
@@ -56,7 +56,41 @@ def verify(args, config):
     app = App(config)
     app.load_datasets()
     app.load_merged_model()
-    app.verify(args.save_results_path)
+    app.verify(args.save_results_file)
+
+
+def test(args, config):
+    app = App(config)
+    app.load_merged_model()
+    app.test(test_dataset=args.test_json_path, save_test_data_path=args.save_results_file)
+
+
+def base_test(args, config):
+    app = App(config)
+    app.load_base_model()
+    app.test(test_dataset=args.test_json_path, save_test_data_path=args.save_results_file)
+
+
+def ollama_verify(args, config):
+    app = App(config)
+    app.load_datasets()
+    app.verify(save_verify_data_path=args.save_results_file, ollama_model=args.model)
+
+
+def ollama_test(args, config):
+    app = App(config)
+    app.test(test_dataset=args.test_json_path, save_test_data_path=args.save_results_file, ollama_model=args.model)
+
+
+def openai_verify(args, config):
+    app = App(config)
+    app.load_datasets()
+    app.verify(save_verify_data_path=args.save_results_file, openai_model=args.model)
+
+
+def openai_test(args, config):
+    app = App(config)
+    app.test(test_dataset=args.test_json_path, save_test_data_path=args.save_results_file, openai_model=args.model)
 
 
 def init_parser(parser):
@@ -88,7 +122,52 @@ def init_parser(parser):
     subparsers.add_parser("run_merged", help="Interact with the merged model").set_defaults(func=run_merged)
     verify_parser = subparsers.add_parser("verify", help="Verify the trained model with the training dataset")
     verify_parser.set_defaults(func=verify)
-    verify_parser.add_argument("--save_results_path", type=str, help="Save verification results here.")
+    verify_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+
+    test_parser = subparsers.add_parser("test", help="Test the trained model with a testing dataset")
+    test_parser.set_defaults(func=test)
+    test_parser.add_argument("--save_results_file", type=str, help="Save test results here.")
+    test_parser.add_argument("test_json_path", type=str, help="Path to test json file.")
+
+    base_parser = subparsers.add_parser("base", help="base subcommands")
+    base_subparser = base_parser.add_subparsers(title="subcommands", description="valid base subcommands")
+    base_run_parser = base_subparser.add_parser("run", help="Interact with the base model.")
+    base_run_parser.set_defaults(func=run_base)
+
+    base_test_parser = base_subparser.add_parser("test", help="Test the base model with the training dataset")
+    base_test_parser.set_defaults(func=base_test)
+    base_test_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+    base_test_parser.add_argument("test_json_path", type=str, help="Path to test json file.")
+
+    ollama_parser = subparsers.add_parser("ollama", help="Ollama subcommands")
+    ollama_subparser = ollama_parser.add_subparsers(title="subcommands", description="valid Ollama subcommands")
+    ollama_verify_parser = ollama_subparser.add_parser(
+        "verify", help="Verify the ollama model with the training dataset"
+    )
+    ollama_verify_parser.set_defaults(func=ollama_verify)
+    ollama_verify_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+    ollama_verify_parser.add_argument("--model", type=str, help="Name of the Ollama model to use.")
+
+    ollama_test_parser = ollama_subparser.add_parser("test", help="Test the Ollama model with the training dataset")
+    ollama_test_parser.set_defaults(func=ollama_test)
+    ollama_test_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+    ollama_test_parser.add_argument("--model", type=str, help="Name of the Ollama model to use.")
+    ollama_test_parser.add_argument("test_json_path", type=str, help="Path to test json file.")
+
+    openai_parser = subparsers.add_parser("openai", help="OpenAI subcommands")
+    openai_subparser = openai_parser.add_subparsers(title="subcommands", description="valid OpenAI subcommands")
+    openai_verify_parser = openai_subparser.add_parser(
+        "verify", help="Verify the OpenAI model with the training dataset"
+    )
+    openai_verify_parser.set_defaults(func=openai_verify)
+    openai_verify_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+    openai_verify_parser.add_argument("--model", type=str, help="Name of the OpenAI model to use.")
+
+    openai_test_parser = openai_subparser.add_parser("test", help="Test the OpenAI model with the training dataset")
+    openai_test_parser.set_defaults(func=openai_test)
+    openai_test_parser.add_argument("--save_results_file", type=str, help="Save verification results here.")
+    openai_test_parser.add_argument("--model", type=str, help="Name of the OpenAI model to use.")
+    openai_test_parser.add_argument("test_json_path", type=str, help="Path to test json file.")
 
 
 def cli(raw_args=None):